requirements.txt
streamlit==1.28.1
python-dotenv==1.0.0
PyPDF2==3.0.1
transformers==4.34.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
langchain==0.0.347
huggingface_hub==0.17.3
requests==2.31.0